+++ routes/posts.php
@@ -15,6 +15,7 @@ $this->domain(config('platform.domain'))->group(function () {
         'prefix'     => \Orchid\Platform\Kernel\Dashboard::prefix('/posts'),
         'namespace'  => 'Orchid\Platform\Http\Controllers\Posts',
     ], function (\Illuminate\Routing\Router $router) {
+
         $router->post('restore/{id?}', [
             'as'   => 'dashboard.posts.restore',
             'uses' => 'PostController@restore',